// client/app/src/generated/dto/video.ts
export interface Video {
    /**
     * @generated from protobuf field: int32 id = 1
     */
    id: number;
    /**
     * @generated from protobuf field: int32 userId = 2
     */
    userId: number;
    /**
     * @generated from protobuf field: string title = 3
     */
    title: string;
    /**
     * @generated from protobuf field: string url = 4
     */
    url: string;
    /**
     * @generated from protobuf field: string objectName = 5
     */
    objectName: string;
    /**
     * @generated from protobuf field: bool enableComment = 6
     */
    enableComment: boolean;
    /**
     * @generated from protobuf field: string visibility = 7
     */
    visibility: string;
    /**
     * @generated from protobuf field: google.protobuf.Timestamp createdAt = 8
     */
    createdAt?: Timestamp;
    /**
     * @generated from protobuf field: string thumbnailUrl = 9
     */
    thumbnailUrl: string;
    /**
     * @generated from protobuf field: google.protobuf.Timestamp postedAt = 10
     */
    postedAt?: Timestamp; // Added
    /**
     * @generated from protobuf field: bool isDraft = 11
     */
    isDraft: boolean; // Added
}
/**
 * @generated from protobuf message VideoFeed
 */
export interface VideoFeed {
    /**
     * @generated from protobuf field: int32 id = 1
     */
    id: number;
    /**
     * @generated from protobuf field: int32 userId = 2
     */
    userId: number;
    /**
     * @generated from protobuf field: string title = 3
     */
    title: string;
    /**
     * @generated from protobuf field: string url = 4
     */
    url: string;
    /**
     * @generated from protobuf field: string objectName = 5
     */
    objectName: string;
    /**
     * @generated from protobuf field: bool enableComment = 6
     */
    enableComment: boolean;
    /**
     * @generated from protobuf field: string visibility = 7
     */
    visibility: string;
    /**
     * @generated from protobuf field: google.protobuf.Timestamp createdAt = 8
     */
    createdAt?: Timestamp;
    /**
     * @generated from protobuf field: int32 commentCount = 9
     */
    commentCount: number;
    /**
     * @generated from protobuf field: int32 likeCount = 10
     */
    likeCount: number;
    /**
     * @generated from protobuf field: int32 viewCount = 11
     */
    viewCount: number;
    /**
     * @generated from protobuf field: string thumbnailUrl = 12
     */
    thumbnailUrl: string;
    /**
     * @generated from protobuf field: google.protobuf.Timestamp postedAt = 13
     */
    postedAt?: Timestamp; // Added
    /**
     * @generated from protobuf field: bool isDraft = 14
     */
    isDraft: boolean; // Added
}
// ...
export const Video = new Video$Type();
// ...
export const VideoFeed = new VideoFeed$Type();